libiot::network

Trait Bind

pub trait Bind {
    type Connection: Connection;
    type Error: Debug;

    // Required method
    fn bind(&mut self, local: &str) -> Result<Self::Connection, Self::Error>;
}

Trait for accepting inbound network connections (server-side).

This trait is implemented by connection types that can listen for and accept incoming connections.

Required Associated Types

type Connection: Connection

The type of connection that will be created for each client

type Error: Debug

Associated error type for bind operations

Required Methods

fn bind(&mut self, local: &str) -> Result<Self::Connection, Self::Error>

Bind to a local address and accept incoming connections.

This is typically a blocking operation that waits for a client to connect.

Arguments

local - Local address to bind to

Returns

Ok(connection) - Incoming connection accepted

Err(e) - Failed to bind or accept connection