libiot::network

Trait Close

pub trait Close {
    type Error: Debug;

    // Required method
    fn close(self) -> Result<(), Self::Error>;
}

Trait for closing network connections.

Provides a clean way to close connections and free associated resources.

Required Associated Types

type Error: Debug

Associated error type for close operations

Required Methods

fn close(self) -> Result<(), Self::Error>

Close the connection and free any associated resources.

After calling this method, the connection should not be used for further operations.

Returns

Ok(()) - Connection closed successfully

Err(e) - Error occurred while closing