libiot::network

Trait Read

pub trait Read {
    type Error: Debug;

    // Required method
    fn read(&mut self, buf: &mut [u8]) -> Result<usize, Self::Error>;
}

Trait for reading data from a network connection.

This trait provides a synchronous interface for reading data from any network connection. It’s designed to be simple and efficient for embedded systems.

Examples

use libiot::network::Read;

fn read_data<R: Read>(reader: &mut R) -> Result<Vec<u8>, R::Error> {
    let mut buffer = [0u8; 1024];
    let bytes_read = reader.read(&mut buffer)?;
    Ok(buffer[..bytes_read].to_vec())
}

Required Associated Types

type Error: Debug

Associated error type for read operations

Required Methods

fn read(&mut self, buf: &mut [u8]) -> Result<usize, Self::Error>

Read data from the connection into the provided buffer.

Returns the number of bytes read. A return value of 0 typically indicates that the connection has been closed by the remote end.

Arguments

buf - Buffer to read data into

Returns

Ok(n) - Number of bytes read

Err(e) - Read error occurred