libiot::network

Trait Write

pub trait Write {
    type Error: Debug;

    // Required methods
    fn write(&mut self, buf: &[u8]) -> Result<usize, Self::Error>;
    fn flush(&mut self) -> Result<(), Self::Error>;
}

Trait for writing data to a network connection.

This trait provides a synchronous interface for writing data to any network connection with support for flushing buffered data.

Examples

use libiot::network::Write;

fn send_message<W: Write>(writer: &mut W, message: &[u8]) -> Result<(), W::Error> {
    writer.write(message)?;
    writer.flush()?;
    Ok(())
}

Required Associated Types

type Error: Debug

Associated error type for write operations

Required Methods

fn write(&mut self, buf: &[u8]) -> Result<usize, Self::Error>

Write data to the connection.

Returns the number of bytes written. The implementation may write fewer bytes than requested.

Arguments

buf - Data to write

Returns

Ok(n) - Number of bytes written

Err(e) - Write error occurred

fn flush(&mut self) -> Result<(), Self::Error>

Flush any buffered write data.

This ensures that all buffered data is sent over the connection. Some implementations may be no-ops if no buffering is used.

Returns

Ok(()) - Flush completed successfully

Err(e) - Flush error occurred