libiot::storage

Trait BlockStorage

pub trait BlockStorage {
    // Required methods
    fn block_size(&self) -> usize;
    fn block_count(&self) -> usize;
}

Block-oriented storage interface.

Many storage devices organize data into fixed-size blocks. This trait provides information about the block structure for devices that use block-based organization.

Examples

use libiot::storage::{Storage, BlockStorage};

fn read_entire_block<S: Storage + BlockStorage>(
    storage: &mut S,
    block_num: usize
) -> Result<Vec<u8>, S::Error> {
    let block_size = storage.block_size();
    let mut buffer = vec![0u8; block_size];
    let offset = (block_num * block_size) as u32;
    storage.read(offset, &mut buffer)?;
    Ok(buffer)
}

Required Methods

fn block_size(&self) -> usize

Get the size of each block in bytes.

Returns

Block size in bytes (commonly 512, 4096, or 8192)

fn block_count(&self) -> usize

Get the total number of blocks on the device.

Returns

Total number of blocks available