libiot::storage

Trait SdMmc

pub trait SdMmc:
    Storage
    + BlockStorage
    + SectorStorage {
    // Required methods
    fn init(&mut self) -> Result<(), <Self as ReadStorage>::Error>;
    fn status(&mut self) -> Result<SdMmcStatus, <Self as ReadStorage>::Error>;
    fn set_block_length(
        &mut self,
        len: u32,
    ) -> Result<(), <Self as ReadStorage>::Error>;
}

SD/MMC card specific operations.

SD and MMC cards require special initialization sequences and have specific status information that can be useful for applications.

Required Methods

fn init(&mut self) -> Result<(), <Self as ReadStorage>::Error>

Initialize the SD/MMC card.

Performs the required initialization sequence for the card, including power-up, command initialization, and capacity detection.

Returns

Ok(()) - Card initialized successfully

Err(error) - Initialization failed

fn status(&mut self) -> Result<SdMmcStatus, <Self as ReadStorage>::Error>

Get the current card status.

Returns detailed status information about the card including capacity, protection status, and other card-specific information.

Returns

Ok(status) - Current card status

Err(error) - Failed to read status

fn set_block_length( &mut self, len: u32, ) -> Result<(), <Self as ReadStorage>::Error>

Set the block length for SDSC cards.

Standard Capacity (SDSC) cards support configurable block lengths. High Capacity (SDHC/SDXC) cards always use 512-byte blocks.

Arguments

len - Block length in bytes (typically 512)

Returns

Ok(()) - Block length set successfully

Err(error) - Failed to set block length