libiot::system::shell

Type Alias CommandFn

pub type CommandFn = fn(argc: usize, argv: &[&str]) -> ShellResult;

Function signature for command handlers.

Command handlers receive the argument count and a slice of argument strings. The first argument (argv[0]) is always the command name itself.

Arguments

argc - Number of arguments (including command name)

argv - Array of argument strings

Returns

A ShellResult indicating success or the type of error that occurred.

Examples

use libiot::system::shell::{ShellResult, CommandFn};

let echo_command: CommandFn = |argc, argv| {
    for i in 1..argc {
        print!("{} ", argv[i]);
    }
    println!();
    ShellResult::Ok
};